move network-related files to new subfolder "network"
