+++ cvmfs/cache.cc
@@ -14,7 +14,7 @@
 #include "compression.h"
 #include "crypto/hash.h"
 #include "directory_entry.h"
-#include "download.h"
+#include "network/download.h"
 #include "quota.h"
 #include "util/posix.h"
 #include "util/smalloc.h"
