+++ cvmfs/cache_posix.cc
@@ -52,9 +52,9 @@
 #include "crypto/hash.h"
 #include "crypto/signature.h"
 #include "directory_entry.h"
-#include "download.h"
 #include "manifest.h"
 #include "manifest_fetch.h"
+#include "network/download.h"
 #include "quota.h"
 #include "shortstring.h"
 #include "statistics.h"
