+++ cvmfs/catalog_diff_tool_impl.h
@@ -10,7 +10,7 @@
 
 #include "catalog.h"
 #include "crypto/hash.h"
-#include "download.h"
+#include "network/download.h"
 #include "util/exception.h"
 #include "util/logging.h"
 #include "util/posix.h"
