+++ cvmfs/cvmfs.cc
@@ -76,7 +76,6 @@
 #include "crypto/hash.h"
 #include "crypto/signature.h"
 #include "directory_entry.h"
-#include "download.h"
 #include "duplex_fuse.h"
 #include "fence.h"
 #include "fetch.h"
@@ -93,6 +92,7 @@
 #include "manifest_fetch.h"
 #include "monitor.h"
 #include "mountpoint.h"
+#include "network/download.h"
 #include "nfs_maps.h"
 #include "notification_client.h"
 #include "options.h"
