+++ cvmfs/fetch.cc
@@ -10,8 +10,8 @@
 #include "backoff.h"
 #include "cache.h"
 #include "clientctx.h"
-#include "download.h"
 #include "interrupt.h"
+#include "network/download.h"
 #include "quota.h"
 #include "statistics.h"
 #include "util/concurrency.h"
