+++ cvmfs/libcvmfs_int.cc
@@ -54,13 +54,13 @@
 #include "crypto/hash.h"
 #include "crypto/signature.h"
 #include "directory_entry.h"
-#include "download.h"
 #include "duplex_sqlite3.h"
 #include "fetch.h"
 #include "globals.h"
 #include "interrupt.h"
 #include "libcvmfs.h"
 #include "lru_md.h"
+#include "network/download.h"
 #include "quota.h"
 #include "shortstring.h"
 #include "sqlitemem.h"
