+++ cvmfs/network/dns.h
@@ -2,8 +2,8 @@
  * This file is part of the CernVM File System.
  */
 
-#ifndef CVMFS_DNS_H_
-#define CVMFS_DNS_H_
+#ifndef CVMFS_NETWORK_DNS_H_
+#define CVMFS_NETWORK_DNS_H_
 
 #include <stdint.h>
 
@@ -410,4 +410,4 @@ class NormalResolver : public Resolver {
 
 }  // namespace dns
 
-#endif  // CVMFS_DNS_H_
+#endif  // CVMFS_NETWORK_DNS_H_
