+++ cvmfs/network/download.cc
