+++ cvmfs/network/s3fanout.cc
