+++ cvmfs/notification_client.h
@@ -7,8 +7,8 @@
 
 #include <string>
 
-#include "download.h"
 #include "fuse_remount.h"
+#include "network/download.h"
 #include "notify/subscriber.h"
 #include "util/pointer.h"
 #include "util/single_copy.h"
