+++ cvmfs/publish/repository.cc
@@ -13,12 +13,12 @@
 #include "catalog_mgr_rw.h"
 #include "crypto/hash.h"
 #include "crypto/signature.h"
-#include "download.h"
 #include "gateway_util.h"
 #include "history_sqlite.h"
 #include "ingestion/ingestion_source.h"
 #include "manifest.h"
 #include "manifest_fetch.h"
+#include "network/download.h"
 #include "publish/except.h"
 #include "publish/repository_util.h"
 #include "publish/settings.h"
