+++ cvmfs/receiver/commit_processor.cc
@@ -13,9 +13,9 @@
 #include "catalog_mgr_ro.h"
 #include "catalog_mgr_rw.h"
 #include "compression.h"
-#include "download.h"
 #include "manifest.h"
 #include "manifest_fetch.h"
+#include "network/download.h"
 #include "params.h"
 #include "signing_tool.h"
 #include "statistics.h"
