+++ cvmfs/swissknife.h
@@ -11,8 +11,8 @@
 #include <vector>
 
 #include "crypto/signature.h"
-#include "download.h"
 #include "manifest_fetch.h"
+#include "network/download.h"
 #include "reflog.h"
 #include "server_tool.h"
 #include "statistics.h"
