+++ cvmfs/swissknife_sync.cc
@@ -39,9 +39,9 @@
 #include "catalog_mgr_ro.h"
 #include "catalog_mgr_rw.h"
 #include "catalog_virtual.h"
-#include "download.h"
 #include "manifest.h"
 #include "monitor.h"
+#include "network/download.h"
 #include "path_filters/dirtab.h"
 #include "reflog.h"
 #include "sanitizer.h"
