+++ cvmfs/talk.cc
@@ -40,13 +40,13 @@
 #include "cache_posix.h"
 #include "catalog_mgr_client.h"
 #include "cvmfs.h"
-#include "download.h"
 #include "duplex_sqlite3.h"
 #include "fuse_remount.h"
 #include "glue_buffer.h"
 #include "loader.h"
 #include "monitor.h"
 #include "mountpoint.h"
+#include "network/download.h"
 #include "nfs_maps.h"
 #include "options.h"
 #include "quota.h"
