+++ cvmfs/upload_s3.cc
@@ -13,8 +13,8 @@
 #include <vector>
 
 #include "compression.h"
+#include "network/s3fanout.h"
 #include "options.h"
-#include "s3fanout.h"
 #include "util/exception.h"
 #include "util/logging.h"
 #include "util/posix.h"
