+++ test/stress/CMakeLists.txt
@@ -3,12 +3,12 @@ set (CVMFS_SOURCE_DIR "${CMAKE_SOURCE_DIR}/cvmfs")
 add_executable(s3benchmark
                test/stress/s3benchmark.cc
                ${CVMFS_SOURCE_DIR}/compression.cc
-               ${CVMFS_SOURCE_DIR}/dns.cc
                ${CVMFS_SOURCE_DIR}/gateway_util.cc
                ${CVMFS_SOURCE_DIR}/json_document.cc
+               ${CVMFS_SOURCE_DIR}/network/dns.cc
+               ${CVMFS_SOURCE_DIR}/network/s3fanout.cc
                ${CVMFS_SOURCE_DIR}/options.cc
                ${CVMFS_SOURCE_DIR}/pack.cc
-               ${CVMFS_SOURCE_DIR}/s3fanout.cc
                ${CVMFS_SOURCE_DIR}/sanitizer.cc
                ${CVMFS_SOURCE_DIR}/session_context.cc
                ${CVMFS_SOURCE_DIR}/ssl.cc
