+++ test/unittests/t_download.cc
@@ -13,9 +13,9 @@
 #include "c_http_server.h"
 #include "compression.h"
 #include "crypto/hash.h"
-#include "download.h"
 #include "interrupt.h"
-#include "sink.h"
+#include "network/download.h"
+#include "network/sink.h"
 #include "statistics.h"
 #include "util/file_guard.h"
 #include "util/posix.h"
