+++ test/unittests/t_fetch.cc
@@ -10,8 +10,8 @@
 #include "backoff.h"
 #include "cache_posix.h"
 #include "crypto/hash.h"
-#include "download.h"
 #include "fetch.h"
+#include "network/download.h"
 #include "statistics.h"
 #include "testutil.h"
 #include "util/atomic.h"
